requirements.txt
python-telegram-bot
apiai
colorama
